Moving settings to TopAdminMenu
