+++ src/Orchard.Web/Modules/Orchard.Azure.MediaServices/Navigation/AdminMenu.cs
@@ -12,13 +12,6 @@ namespace Orchard.Azure.MediaServices.Navigation {
                 .Add(T("Media"), "6", menu => menu
                     .Add(T("Microsoft Azure Media Jobs"), "10.0", item => item.Action("Index", "Job", new { area = "Orchard.Azure.MediaServices" })
                         .Permission(Permissions.ManageCloudMediaJobs)));
-
-            builder
-                .Add(T("Settings"), menu => menu
-                    .Add(T("Microsoft Azure Media"), "10.0", item => item.Action("Index", "Settings", new { area = "Orchard.Azure.MediaServices" })
-                        .Permission(StandardPermissions.SiteOwner)
-                        .Permission(Permissions.ManageCloudMediaSettings)
-                ));
         }
     }
 }
