+++ src/Orchard.Web/Modules/Orchard.Azure.MediaServices/Orchard.Azure.MediaServices.csproj
@@ -151,6 +151,7 @@
     <Content Include="Content\MSAdaptiveStreamingPlugin.swf" />
     <Content Include="Content\StrobeMediaPlayback.swf" />
     <Compile Include="Models\EncodingPreset.cs" />
+    <Compile Include="Navigation\TopAdminMenu.cs" />
     <Compile Include="Shapes\CloudVideoPlayerShape.cs" />
     <Compile Include="Helpers\BitExtensions.cs" />
     <Compile Include="Infrastructure\Assets\AssetDriverResult.cs" />
