+++ src/Orchard.Web/Modules/Orchard.Packaging/Orchard.Packaging.csproj
@@ -75,8 +75,10 @@
     <Reference Include="System.Xml.Linq" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="AdminMenu.cs" />
-    <Compile Include="AdminMenuGalleryUpdates.cs" />
+    <Compile Include="Helpers\NavigationHelpers.cs" />
+    <Compile Include="Navigation\TopAdminMenu.cs" />
+    <Compile Include="Navigation\AdminMenu.cs" />
+    <Compile Include="Navigation\AdminMenuGalleryUpdates.cs" />
     <Compile Include="Commands\PackagingCommands.cs" />
     <Compile Include="Controllers\GalleryUpdatesController.cs" />
     <Compile Include="Controllers\PackagingServicesController.cs" />
